server/requirements.txt
flask>=2.2.0
flask-cors>=3.0.0
transformers>=4.35.0
torch
numpy>=1.24.0
nltk>=3.8.0
gunicorn>=20.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
newspaper3k>=0.2.8